+++ graflo/db/manager.py
@@ -29,6 +29,7 @@ from graflo.db.arango.conn import ArangoConnection
 from graflo.db.connection import DBConfig, TARGET_DATABASES
 from graflo.onto import DBType
 from graflo.db.falkordb.conn import FalkordbConnection
+from graflo.db.graflo_backend.connection import GraFloBackendConnection
 from graflo.db.memgraph.conn import MemgraphConnection
 from graflo.db.nebula.conn import NebulaConnection
 from graflo.db.neo4j.conn import Neo4jConnection
@@ -63,6 +64,7 @@ class ConnectionManager:
         DBType.MEMGRAPH: MemgraphConnection,
         DBType.NEBULA: NebulaConnection,
         DBType.POSTGRES: PostgresConnection,
+        DBType.GRAFLO_BACKEND: GraFloBackendConnection,
     }
 
     @classmethod